100% in 4 testers

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: mlazzare <mlazzare@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/11 10:46:12 by mlazzare          #+#    #+#             */
-/*   Updated: 2021/03/28 15:36:19 by mlazzare         ###   ########.fr       */
+/*   Updated: 2021/03/31 13:00:28 by mlazzare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ typedef	struct		s_list
 }					t_list;
 
 size_t			ft_strlen(const char *s);
-int 				ft_numlen_base(long long n, int base);
+int 				ft_numlen_base(unsigned long n, int base);
 int					ft_putnbr_base(unsigned long long n, char *base);
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
